move target utils to own file
