+++ frontend/src/metabase/meta/Card.js
@@ -1,10 +1,10 @@
 import {
   getTemplateTagParameters,
-  getParameterTargetField,
   parameterToMBQLFilter,
   normalizeParameterValue,
   getValuePopulatedParameters,
 } from "metabase/meta/Parameter";
+import { getParameterTargetField } from "metabase/parameters/utils/targets";
 
 import * as Query from "metabase/lib/query/query";
 import * as Q_DEPRECATED from "metabase/lib/query"; // legacy
