+++ frontend/src/metabase/meta/Dashboard.js
@@ -7,7 +7,7 @@ import type {
   ParameterOption,
   Parameter,
 } from "metabase-types/types/Parameter";
-import { getParameterTargetField } from "metabase/meta/Parameter";
+import { getParameterTargetField } from "metabase/parameters/utils/targets";
 import { slugify } from "metabase/lib/formatting";
 
 export type ParameterSection = {
