+++ frontend/test/metabase/meta/Parameter.unit.spec.js
@@ -1,6 +1,5 @@
 import MetabaseSettings from "metabase/lib/settings";
 import {
-  getParameterTargetField,
   dateParameterValueToMBQL,
   stringParameterValueToMBQL,
   numberParameterValueToMBQL,
@@ -14,11 +13,7 @@ import {
   buildHiddenParametersSlugSet,
   getVisibleParameters,
 } from "metabase/meta/Parameter";
-import {
-  metadata,
-  PRODUCTS,
-  SAMPLE_DATASET,
-} from "__support__/sample_dataset_fixture";
+import { metadata, PRODUCTS } from "__support__/sample_dataset_fixture";
 
 MetabaseSettings.get = jest.fn();
 
@@ -299,39 +294,6 @@ describe("metabase/meta/Parameter", () => {
           metadata,
         ),
       ).toEqual(["between", ["field", PRODUCTS.RATING.id, null], 1, 100]);
-    });
-  });
-
-  describe("getParameterTargetField", () => {
-    it("should return null when the target is not a dimension", () => {
-      expect(getParameterTargetField(["variable", "foo"], metadata)).toBe(null);
-    });
-
-    it("should return the mapped field behind a template tag field filter", () => {
-      const target = ["dimension", ["template-tag", "foo"]];
-      const question = SAMPLE_DATASET.nativeQuestion({
-        query: "select * from PRODUCTS where {{foo}}",
-        "template-tags": {
-          foo: {
-            type: "dimension",
-            dimension: ["field", PRODUCTS.CATEGORY.id, null],
-          },
-        },
-      });
-
-      expect(getParameterTargetField(target, metadata, question)).toBe(
-        PRODUCTS.CATEGORY,
-      );
-    });
-
-    it("should return the target field", () => {
-      const target = ["dimension", ["field", PRODUCTS.CATEGORY.id, null]];
-      const question = SAMPLE_DATASET.question({
-        "source-table": PRODUCTS.id,
-      });
-      expect(getParameterTargetField(target, metadata, question)).toBe(
-        PRODUCTS.CATEGORY,
-      );
     });
   });
